Actualizacion cosas

+++ tienda.sql
@@ -5,7 +5,7 @@ USE tienda;
 CREATE TABLE productos (
   id BIGINT AUTO_INCREMENT PRIMARY KEY,
   nombre VARCHAR(255) NOT NULL,
-  precio DECIMAL(10, 2) NOT NULL,
+  precio DOUBLE NOT NULL,
   cantidad INT NOT NULL,
   categoria VARCHAR(255)
 );
